Fix fullstack defects: README run command, decimal bounds, DDL FK/CHECK import
- README told users to run ./mvnw, which the common catalog deletes; use mvn.
- min/max are BigDecimal end to end: decimal bounds render as @DecimalMin/@DecimalMax
  and step="any" on BigDecimal fields; integral fields reject fractional bounds.
- DDL import turns single-column foreign keys to imported single-key tables into
  MANY_TO_ONE relations (column dropped) and CHECK (col IN (...)) into ENUM fields;
  the import wire carries relations.
- Enum-constant keyword check is exact-case (NEW/DEFAULT are ordinary constants).

+++ src/main/java/com/menora/initializr/fullstack/FieldDefinition.java
@@ -1,12 +1,14 @@
 package com.menora.initializr.fullstack;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
  * One field on a user-defined entity. Submitted by clients as JSON; converted from
  * the wire form by {@link FullstackRequestValidator}.
  *
- * <p>{@code min}/{@code max} are numeric bounds (integral; rendered as Bean Validation
+ * <p>{@code min}/{@code max} are numeric bounds (whole numbers on integral types, any
+ * decimal on {@code BigDecimal}; rendered as Bean Validation
  * {@code @Min}/{@code @Max} on integral types or {@code @DecimalMin}/{@code @DecimalMax}
  * on {@code BigDecimal}). {@code pattern} is a regex and {@code email} a convenience flag,
  * both for {@code STRING} fields ({@code @Pattern} / {@code @Email}).
@@ -37,8 +39,8 @@ public record FieldDefinition(
         boolean required,
         boolean unique,
         Integer length,
-        Long min,
-        Long max,
+        BigDecimal min,
+        BigDecimal max,
         String pattern,
         boolean email,
         List<String> enumValues,
@@ -51,7 +53,7 @@ public record FieldDefinition(
     /** Back-compat constructor without the {@code label}/{@code readOnly}/{@code defaultValue}
      *  per-field props (no custom label, editable, no default). Keeps existing callers/tests compiling. */
     public FieldDefinition(String name, FieldType type, boolean primaryKey, boolean generated,
-                           boolean required, boolean unique, Integer length, Long min, Long max,
+                           boolean required, boolean unique, Integer length, BigDecimal min, BigDecimal max,
                            String pattern, boolean email, List<String> enumValues,
                            boolean searchable, boolean filterable) {
         this(name, type, primaryKey, generated, required, unique, length, min, max,
@@ -60,7 +62,7 @@ public record FieldDefinition(
 
     /** Back-compat constructor without {@code defaultValue} (no default). */
     public FieldDefinition(String name, FieldType type, boolean primaryKey, boolean generated,
-                           boolean required, boolean unique, Integer length, Long min, Long max,
+                           boolean required, boolean unique, Integer length, BigDecimal min, BigDecimal max,
                            String pattern, boolean email, List<String> enumValues,
                            boolean searchable, boolean filterable, String label, boolean readOnly) {
         this(name, type, primaryKey, generated, required, unique, length, min, max,
